Support for E01 files
First version of Ewf format. These are files used for taking copies of
media typically in the digital forensics arena. These files are not
meant to be edited once they are created. A variety of hashes are built
into the format so that changes can be detected.

Basic testing complete. MD5 of image read through
this code matches MD5 reported by EnCase/FTK Imager.