Read/Write support for the remaining NFS classes (#85)
A continuation of #83, this adds read/write support to the the other NFS classes.
Basic unit tests have been added, they make sure the data roundtrips correctly when serialized.

Adds a missing `Nfs3StableHow` enum.

It also marks all of the classes used by RPC calls in NFS as public, this allows a you to implement a NFS server on top of DiscUtils. The constructors which take an `XdrDataReader` and `Write` methods are kept internal and the classes are sealed.