Improve DiscUtils extensibility
Currently, it's difficult to build and maintain out-of-tree extensions to DiscUtils. Some classes which you need to implement are either `internal`, or have (abstract) methods or properties which are `internal`. This makes it impossible to implement these classes in a 3rd party project.

This commit improves DiscUtils extensibility a bit by:
- Making `VirtualDiskFactory`, `VirtualDiskFactoryAttribute` public
- Making `FileLocator` public
- Making `PartitionInfo.VolumeType` public
- Making `VirtualDiskLayer.Capacity` and `VirtualDiskLayer.RelativeFileLocator` public